openAIFunctions to separate file
